Squash merged 'model-update', major refactor
The GWT client, related code and dependencies have been removed, UI is now entirely
based on Typescript and Lit element. A major refactor of the asset model has been
completed that facilitates a strong type system in java code, custom Value types
have been removed and instead standard POJOs are used with jackson serialisation/
deserialisation. Assets now require a POJO for each asset type and JPA single table
inheritance has been implemented for assets, also Hibernate types library is being
utilised for marshalling between JSON and POJOs. The protocol SPI has been refactored
moving from a singleton model to an instance per protocol configuration (now known as
agent). Whereas previously an agent asset could have multiple protocol configuration
attributes where each represented a protocol instance, now an agent represents a
protocol instance and the old protocol configuration is replaced with agent attribute.
The agent class is responsible for instantiating the associated protocol instance when
the agent is started. Attribute's linked to protocols now define all of their configuration
within the AgentLink meta item; each protocol can have its' own sub type of AgentLink
that defines the available configuration items for linked attributes.

Where possible java dependencies have been updated to latest stable version.

Summary

- GWT client removed
- Java dependencies updated
- Model overhaul
- Protocol SPI change

+++ manager/src/main/java/org/openremote/manager/rules/RulesClock.java
@@ -32,19 +32,19 @@ import java.util.function.Predicate;
  */
 public class RulesClock {
 
-    final public double timestamp;
+    final public long timestamp;
     final public LocalDateTime time;
 
     public RulesClock(TimerService timerService) {
         this(timerService.getCurrentTimeMillis());
     }
 
-    public RulesClock(double timestamp) {
+    public RulesClock(long timestamp) {
         this.timestamp = timestamp;
-        time = LocalDateTime.ofInstant(Instant.ofEpochMilli((long) this.timestamp), ZoneId.systemDefault());
+        time = LocalDateTime.ofInstant(Instant.ofEpochMilli(this.timestamp), ZoneId.systemDefault());
     }
 
-    public double getTimestamp() {
+    public long getTimestamp() {
         return timestamp;
     }
